requirements.txt
openai==1.11.0
streamlit==1.32.2
langchain==0.1.14
langchain-community==0.0.38
python-dotenv==1.0.1
httpx==0.27.0
langchain-openai==0.0.6
beautifulsoup4==4-4.13.4